Remove gRPC-based API implementation
