+++ current_results/Dockerfile
@@ -19,4 +19,3 @@ COPY --from=build /app/bin/server /app/bin/
 # Start server.
 EXPOSE 8080
 ENTRYPOINT ["/app/bin/server"]
-CMD ["--server=grpc"]
