+++ current_results/README.md
@@ -8,10 +8,8 @@ This service is the backend for https://dart-current-results.web.app/
 which provides a filterable view of the current test status across
 all configurations.
 
-The current results service is a gRPC server written in Dart, running
-on Cloud Run. Cloud Endpoints is used to configure access to the server's API,
-and the ESPv2 proxy is used to manage the API and provide a REST translation of
-the API.
+The current results service is a REST server written in Dart, running
+on Cloud Run.
 
 The results service listens for new Dart CI results uploaded to cloud storage,
 and loads them into an in-memory database of all current CI results.
@@ -38,11 +36,9 @@ To generate the required Dart files for the protos, run
 `tools/generate_protogen.sh`.
 
 ### Deployment
-There are three deployment scripts, one for the initial deployment of the
+There are two deployment scripts, one for the initial configuration of the
 current results service to a cloud project, and one for redeploying the
-current results server, and one for redeploying the Cloud Endpoints
-configuration and the ESPv2 proxy. Three deployment scripts, there
-are three deployment scripts.
+current results server.
 
 These scripts should be run with the current directory set to the
 current_results directory in a checkout of the dart-lang/dart_ci
@@ -50,7 +46,6 @@ repository, with the version you wish to deploy checked out.
 
 * initial_configuration.sh (only needed once to setup a new project)
 * deploy_service.sh
-* deploy_endpoints_configuration.sh
 
 ## Services
 
@@ -70,6 +65,6 @@ updated from that build.
 
 ### Serving results data
 
-A gRPC service is defined, with a query message type and a response message
+A REST service is defined, with a query message type and a response message
 type defined by protocol buffers.  It allows test result queries by test name or
 test name prefix and by configuration or configuration prefix.
