--- current_results/endpoints/cloud_run_api_config_template.yaml
@@ -1,25 +1,0 @@
-# Copyright (c) 2020, the Dart project authors.  Please see the AUTHORS file
-# for details. All rights reserved. Use of this source code is governed by a
-# BSD-style license that can be found in the LICENSE file.
-
-# The configuration schema is defined by service.proto file
-# https://github.com/googleapis/googleapis/blob/master/google/api/service.proto
-type: google.api.Service
-config_version: 3
-
-# Name of the service configuration.
-name: SERVICE_HOSTNAME
-
-# API title to appear in the user interface (Google Cloud Console).
-title: Current Results API
-apis:
-- name: current_results.Query
-usage:
-  rules:
-  - selector: "*"
-    allow_unregistered_calls: true
-backend:
-  rules:
-    - selector: "*"
-      address: grpcs://BACKEND_HOSTNAME
-      deadline: 60.0
